Rapid ICU Patient Data Review System
Review a patient's data in the ICU can be incredibly cumbersome,
especially for pre-rounding. The current EMR at NYP has greatly
improved patient care, avoided errors and made things more efficient
for physicians as well as nursing staff. However, the time required to
understand a patient's chart still is very time consuming. In some
cases such as cardiac arrests or rapid responses learning the
information rapidly is of utmost importance to saving the patient's
life.
The goal of this tool is reduce the amount of clicking and searching an
Intensivist has to do make clinical decisions on a daily, often hourly
basis.